native: 100 twin comparisons had silently lost their sign
Every fixed-VA global is typed `unsigned int` - by alias_globals.py for the
native build, and by verify_coexec's gdef for the harness. That is fine for the
bit-twiddling the engine mostly does, but it INVERTS a signed test:

    g_x < 0     never fires
    0 < g_x     collapses to "g_x != 0"
    -1 < g_x    becomes "g_x >= 0xffffffff", i.e. almost never true

and it is invisible. It compiles, and co-exec keeps passing, because BOTH sides
are wrong together: the harness types the twin's globals the same way the
native build does, so a seeded run only catches it if some seed actually drives
the value negative. Nothing did.

Found while seeding GeoTransformDispatchAndApply, whose original is `jle`, then
`je`+`jl` (together "greater than zero"). Three of six seeded cases mismatched
the moment a negative went in, and the dispatch index came out wrong.

tools/decomp/audit_signed.py (make signed-audit) generalises it: match the
comparison shapes in every NON_MATCHING body, then cross-check each hit against
the ORIGINAL's own jcc opcodes, dropping any function that branches only
unsigned. 100 hits, 0 dropped - which stands to reason, since Ghidra only wrote
`< 0` where the original tested the sign in the first place. `--fix` inserts
the casts; it skips comment text, because the harness scans comments for global
names and a `g_x` in prose becomes an "unmapped global" SKIP.

100 casts across 46 files. Four of them are in the render path that is already
live in native-full - tristrip_batch_emit, tristrip_batch_emit2,
tristrip_batch_emit3_cap, textured_tri_rasterize_alpha - where the guard is
`0 < g_min_007af98c` on a min-coordinate accumulator that can legitimately go
negative.

GeoTransformDispatchAndApply also had a dropped store: 0x54204c is written
TWICE, the dispatch INDEX at 0x489891 and then the loaded target at 0x48989d.
Ghidra kept only the second. It is observable, so it is back.

The frame is unchanged at 62300 px - no seeded scene drove those accumulators
negative, which is exactly why this survived until now.

Gates: matching byte-identical; native-char-check byte-identical to the
verified bundle; native-frame-check 62300 px; both wasm32 parity gates
byte-identical; verify_emit 14/14, verify_project 30/30, verify_mesh 10/10,
verify_rot 18/18, frame-core-check 51/51.

+++ src/audio/audio_mixer_knob_update.c
@@ -129,7 +129,7 @@ void AudioMixerKnobUpdate(void)
   int iVar2;
   
   if (g_audioMixerKnob2 == 0) {
-    if (((g_audioStateDisp50b4 & 4) != 0) && (0 < g_counter_0054359c)) {
+    if (((g_audioStateDisp50b4 & 4) != 0) && (0 < (int)g_counter_0054359c)) {
       SetJmp_Push16Call_004a1ad0();
       g_counter_0054359c = g_counter_0054359c + -1;
     }
@@ -164,7 +164,7 @@ void AudioMixerKnobUpdate(void)
     ScaledChainStore24();
   }
   if (g_audioMixerKnob == 0) {
-    if (((g_audioStateDisp50b4 & 0x400) != 0) && (0 < g_counter_005433c8)) {
+    if (((g_audioStateDisp50b4 & 0x400) != 0) && (0 < (int)g_counter_005433c8)) {
       SetJmp_Push16Call_004a1ad0();
       g_counter_005433c8 = g_counter_005433c8 + -1;
     }
