+++ src/audio/linked_list_builder.c
@@ -148,7 +148,7 @@ void LinkedListBuilder(void)
     *MK4_NODE(int, (iVar1 + 1 + iVar4)) = iVar3;
     g_xformLoopCounter = MK4_NODE_AT(int, iVar3, 8);
     if (1 < g_xformLoopCounter) {
-      for (g_xformLoopCounter = g_xformLoopCounter + -2; -1 < g_xformLoopCounter; g_xformLoopCounter = g_xformLoopCounter + -1) {
+      for (g_xformLoopCounter = g_xformLoopCounter + -2; -1 < (int)g_xformLoopCounter; g_xformLoopCounter = g_xformLoopCounter + -1) {
         iVar5 = iVar4 * 4;
         iVar4 = iVar4 + iVar2;
         *MK4_NODE(int, iVar5 + 4 + iVar1) = iVar3;
